requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
django-environ==0.11.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
wheel==0.44.0